requirements.txt
Flask
Flask-Login
Flask-Bcrypt
Flask-Migrate
Flask-WTF
python-dotenv
celery
redis
torch
openai-whisper
openai
google-generativeai
yt-dlp
pydub
gunicorn
gevent
click
requests
konlpy
google-cloud-texttospeech